backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.53
langchain-ollama==0.2.1
langfuse==2.58.2
psycopg[binary]==3.2.3
python-dotenv==1.0.1
numpy==2.1.2
pandas==2.2.3
APScheduler==3.10.4
